- Refactored Metadata Presets to Use .m_preset files #119 #111
This changes to code so presets are single files.  There are default presets in the plugin dir to start with, and then each project gets these plugin presets.  Then changes to presets are saved in the project, to allow reordering, editing, adding/removing.

This structure will be used for Proxy presets (ffmpeg) as well.