Refactor Image Sequence Tiles for Performance #175
Originally when image sequences were loaded, it will do all the calcs (and thumbnail) for every image in the sequence, and store all the data (and thumbnail) in the sequence tile - this was slow for large seqs.

Now only the data and thumb for the first file is done, and the filepaths for the seq files are stored in the Sequence Tile.